modifications general and user feature

+++ AdaTranslation.Application/Languages/Dtos/LanguageDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace AdaTranslation.Application.Languages.Dtos
 {
-    public record LanguageDto( long Id,string Description);
+    public record LanguageDto( int Id,string Description);
 }
